requirements.txt
# ── Runtime dependencies ──────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic[email]>=2.7.0
pydantic-settings>=2.3.0

# Async MySQL driver
aiomysql>=0.2.0

# Required for sha256_password / caching_sha2_password MySQL auth (MySQL 8+)
cryptography>=42.0.0

# Sync MySQL driver (Alembic migrations)
pymysql>=1.1.0

# ORM
sqlalchemy>=2.0.35

# Auth: password hashing + JWT
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0

# Required by SQLAlchemy async/ORM internals
greenlet>=3.0.0

# ASGI server process manager (production)
gunicorn>=22.0.0
